Settings: Move into config folder
